+++ lib/serega/plugins/explicit_many_option/validations/check_opt_many.rb
@@ -10,23 +10,25 @@ class Serega
         class << self
           #
           # Checks attribute :many option must be provided with relations
+          # (attributes with the :serializer option or a block defining
+          # a nested serializer)
           #
           # @param opts [Hash] Attribute options
+          # @param block [nil, Proc] Attribute block
           #
           # @raise [SeregaError] Attribute validation error
           #
           # @return [void]
           #
-          def call(opts)
-            serializer = opts[:serializer]
-            return unless serializer
+          def call(opts, block = nil)
+            return if !opts[:serializer] && !block
 
             many_option_exists = opts.key?(:many)
             return if many_option_exists
 
             raise SeregaError,
               "Attribute option :many [Boolean] must be provided" \
-              " for attributes with :serializer option"
+              " for attributes with :serializer option or a block"
           end
         end
       end