+++ lib/serega/validations/attribute/check_opt_delegate.rb
@@ -18,11 +18,11 @@ class Serega
           #
           # @return [void]
           #
-          def call(opts, block = nil)
+          def call(opts)
             return unless opts.key?(:delegate)
 
             check_opt_delegate(opts)
-            check_usage_with_other_params(opts, block)
+            check_usage_with_other_params(opts)
           end
 
           private
@@ -56,12 +56,11 @@ class Serega
             Utils::CheckAllowedKeys.call(delegate_opts, %i[to method allow_nil], :delegate)
           end
 
-          def check_usage_with_other_params(opts, block)
+          def check_usage_with_other_params(opts)
             raise SeregaError, "Option :delegate can not be used together with option :method" if opts.key?(:method)
             raise SeregaError, "Option :delegate can not be used together with option :const" if opts.key?(:const)
             raise SeregaError, "Option :delegate can not be used together with option :value" if opts.key?(:value)
             raise SeregaError, "Option :delegate can not be used together with option :batch" if opts.key?(:batch)
-            raise SeregaError, "Option :delegate can not be used together with block" if block
           end
         end
       end